Moved all JFR test cases.
