+++ substratevm/src/com.oracle.svm.test.jdk11/src/com/oracle/svm/test/jdk11/jfrtest/ClassEvent.java
@@ -24,7 +24,7 @@
  * questions.
  */
 
-package com.oracle.svm.jfrtest;
+package com.oracle.svm.test.jdk11.jfrtest;
 
 import jdk.jfr.Description;
 import jdk.jfr.Event;
