+++ substratevm/src/com.oracle.svm.test.jdk11/src/com/oracle/svm/test/jdk11/jfrtest/TestClassEvent.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2020, 2021, Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2020, 2021, Red Hat Inc. All rights reserved.
+ * Copyright (c) 2021, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2021, Red Hat Inc. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,31 +24,31 @@
  * questions.
  */
 
-package com.oracle.svm.jfrtest;
+package com.oracle.svm.test.jdk11.jfrtest;
 
 import jdk.jfr.Recording;
-import jdk.jfr.consumer.RecordingFile;
 
 import static org.junit.Assert.assertNotNull;
+
+import jdk.jfr.consumer.RecordingFile;
 import org.junit.Test;
 
-public class TestRecordingParsable {
-
+public class TestClassEvent {
     @Test
     public void test() throws Exception {
         JFR jfr = new LocalJFR();
-        Recording recording = jfr.startRecording("TestRecordingParsable");
+        Recording recording = jfr.startRecording("TestSingleEvent");
 
-        StringEvent event = new StringEvent();
-        event.message = "Event has been generated!";
+        ClassEvent event = new ClassEvent();
+        event.clazz = TestClassEvent.class;
         event.commit();
 
         jfr.endRecording(recording);
-
         try (RecordingFile recordingFile = new RecordingFile(recording.getDestination())) {
             assertNotNull(recordingFile);
         } finally {
             jfr.cleanupRecording(recording);
         }
     }
+
 }
